Increase the build timeout
The build subtask consists of a lot of things:

* create a chroot, by extracting an archive containing a prepared
  minimal chroot
* update the minimal chroot
* install the build dependencies (from apt) in the chroot
* create a virtualenv in the chroot
* install the dependencies (from pip) in the virtualenv, compiling some
  of them
* install ideascube itself in the virtualenv
* make a .deb package of the virtualenv

All this can take a bit of time.

Right now, it commonly reaches the 20 minutes timeout, because we're
having bad IO (either network or disk) on the build VM.

Setting the timeout to 40 minutes is not really a fix, but it is
necessary for now, until we can get the build times down to a reasonable
number.