workers: Take build deps from the Jessie backports
The --extra-repository will add the Jessie backports as a valid archive
to use in the chroot.

But unfortunately, that's not enough because the Release file for the
backports archive has "NotAutomatic: yes". As a result, even though the
backports contain the dependency we need, it doesn't get taken from
there automatically.

That's where the --build-dep-resolver=aptitude comes into play: aptitude
also stumbles on the missing dependency like apt, but then it finds it
in the backports archive and automatically decides to take it (and only
it) from there.

http://unix.stackexchange.com/a/298116/178537

Fixes #8