properly call the sbuild command
Running the sbuild command in a bash-like shell, we should have run it
as follows:

    sbuild ... --extra-repository='deb http://foo jessie-backports main'

That is because without the quotes, bash would have considered the
spaces as separating multiple parameters, and would have passed them
separately to the sbuild command.

But we're not running sbuild in a bash-like shell here, we're running it
through Python's subprocess module, which does its own space escaping to
properly pass parameters to the command being run.

As a result, we do not need those quotes, we only need to pass all of it
as a single argument, that is:

    ['sbuild', ..., '--extra-repository=deb http://foo jessie-backports main']

The equivalent of bash's splitting on space without the quotes would be
this in Python:

    ['sbuild', ..., '--extra-repository=deb', 'http://foo', 'jessie-backports', 'main']

So, since Python does its own space escaping, adding the quotes as we
were doing would get Python to actually pass those quotes to sbuild,
which would in turn add the quotes in the sources.list file, which is
obviously incorrect.